Moved SingleMetadataReults to shared components, added params for showing category and municipality.
